Update audit.js and fix error catching logic in should().throw()
Fetch/update audit.js in non-WPT layout tests and also fix the error
in the error catching logic in should().throw() assertion.

The updated code checks if the |error| object passed by try/catch
is "DOMException". This is unnecessarily specific, so this CL changes
the check to look for the "Error" type, which is more generic.

Bug: 865614
Test: All layout tests pass.
Change-Id: I16acacb26c194a0ff950aca05e931195bf55167f