Add observable constraints for getDisplayMedia()
This CL adds observable constraints to MediaStream objects returned by
getDisplayMedia() calls. As defiend in the spec, these constraints are set not
before the stream is returned based on user choice.

Information about these constraints are stored in DisplayMediaInformation
struct. It is set in the low level device selection and then parsed by
UserMediaProcessor before passing it to blink.

Bug: 326740
Change-Id: I81d5f2eb24d2c1a37df0fbe2d34bc469fa575bf1