Avoid overflowing snap areas in scrollTo wpt tests
scrollTo-scrollBy-snaps.html is failing on wpt dashboard. This is
caused by the uncertainty of the window size for testharness tests.
If the window size becomes small enough that the snap area covers the
snapport, the snap won't happen. This test changes the landing position
of scrollTo and scrollBy so that the snap area won't cover ths snapport
regardless of the window size.

More specifically, the snap area's vertical range is 1000-1600, and the
snap positin is 1000. If we land on 1200 and the window's height is
smaller than 400, the snap area will cover the window. But if we land
on 900, the snap area won't cover the window regardless of the height.

Bug: 876052
Change-Id: I448ecb1e3cc2cc458d467472ca38d49bf01d2901